requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
Pillow == 9.1.0
psycopg2-binary
git+https://github.com/eternnoir/pyTelegramBotAPI.git